Move pmatrix to its own file
